Add PostgreSQL migration files for DB resource groups
