+++ plugin/trino-resource-group-managers/src/main/java/io/trino/plugin/resourcegroups/db/FlywayMigration.java
@@ -17,6 +17,8 @@ import io.airlift.log.Logger;
 import org.flywaydb.core.Flyway;
 import org.flywaydb.core.api.output.MigrateResult;
 
+import static java.lang.String.format;
+
 public class FlywayMigration
 {
     private static final Logger log = Logger.get(FlywayMigration.class);
@@ -25,12 +27,28 @@ public class FlywayMigration
     {
     }
 
+    private static String getLocation(String configDbUrl)
+    {
+        if (configDbUrl.startsWith("jdbc:postgresql")) {
+            return "/db/migration/postgresql";
+        }
+        else if (configDbUrl.startsWith("jdbc:oracle")) {
+            return "/db/migration/oracle";
+        }
+        else if (configDbUrl.startsWith("jdbc:mysql")) {
+            return "/db/migration/mysql";
+        }
+        // validation is not performed in DbResourceGroupConfig because DB backed
+        // resource group tests use the h2 database.
+        throw new IllegalArgumentException(format("Invalid JDBC URL: %s. Only PostgreSQL, MySQL, and Oracle are supported.", configDbUrl));
+    }
+
     public static void migrate(DbResourceGroupConfig config)
     {
         log.info("Performing migrations...");
         Flyway flyway = Flyway.configure()
                 .dataSource(config.getConfigDbUrl(), config.getConfigDbUser(), config.getConfigDbPassword())
-                .locations("/db/migration/mysql")
+                .locations(getLocation(config.getConfigDbUrl()))
                 .baselineOnMigrate(true)
                 .baselineVersion("0")
                 .load();
