+++ plugin/trino-resource-group-managers/src/main/resources/db/migration/postgresql/V1__add_resource_groups_global_properties.sql
@@ -1,0 +1,5 @@
+CREATE TABLE IF NOT EXISTS resource_groups_global_properties (
+    name VARCHAR(128) NOT NULL PRIMARY KEY,
+    value VARCHAR(512) NULL,
+    CHECK (name in ('cpu_quota_period'))
+);
