+++ plugin/trino-resource-group-managers/src/main/resources/db/migration/postgresql/V3__add_selectors.sql
@@ -1,0 +1,10 @@
+CREATE TABLE IF NOT EXISTS selectors (
+     resource_group_id BIGINT NOT NULL,
+     priority BIGINT NOT NULL,
+     user_regex VARCHAR(512),
+     source_regex VARCHAR(512),
+     query_type VARCHAR(512),
+     client_tags VARCHAR(512),
+     selector_resource_estimate VARCHAR(1024),
+     FOREIGN KEY (resource_group_id) REFERENCES resource_groups (resource_group_id) ON DELETE CASCADE
+);
